pyproject.toml
[project]
name = "agenticdiscordbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "discord.py~=2.5",
    "mcrcon~=0.7",
    "python-dotenv~=1.1",
    "loguru~=0.7",
    "pydantic-ai[a2a,evals,fastmcp,google,logfire,mcp,openai]~=1.0",
    "scrapling[all]~=0.3",
    "beautifulsoup4~=4.0",
    "lxml~=6.0"
]
